feat(AI): 添加关闭AI思考过程输出的功能
在配置中新增no_think选项，当设置为true时，AI将直接输出结果而不显示思考过程
该功能适用于生产环境，可减少不必要的输出并提高响应速度
```

这个提交消息遵循了以下原则：
1. 使用feat类型，因为这是一个新功能
2. 添加了scope(AI)以明确功能范围
3. 描述简洁(不超过50字符)且使用祈使语气
4. 在正文中补充说明了功能的目的和适用场景
5. 使用中文简洁清晰地描述了变更内容
