+++ pkg/ai/openai.go
@@ -1,14 +1,99 @@
-/*
-Copyright 2023 The K8sGPT Authors.
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-    http://www.apache.org/licenses/LICENSE-2.0
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
 package ai
+
+import (
+	"errors"
+	"net/http"
+	"net/url"
+
+	"github.com/sashabaranov/go-openai"
+)
+
+const openAIClientName = "openai"
+
+type OpenAIClient struct {
+	nopCloser
+	client      *openai.Client
+	model       string
+	temperature float32
+	topP        float32
+	tools       []openai.Tool
+	maxHistory  int32
+	memory      *memoryService
+
+	// organizationId string
+}
+
+func (c *OpenAIClient) GetName() string {
+	return openAIClientName
+}
+
+// OpenAIHeaderTransport is an http.RoundTripper that adds the given headers to each request.
+type OpenAIHeaderTransport struct {
+	Origin  http.RoundTripper
+	Headers []http.Header
+}
+
+// RoundTrip implements the http.RoundTripper interface.
+func (t *OpenAIHeaderTransport) RoundTrip(req *http.Request) (*http.Response, error) {
+	// Clone the request to avoid modifying the original request
+	clonedReq := req.Clone(req.Context())
+	for _, header := range t.Headers {
+		for key, values := range header {
+			// Possible values per header:  RFC 2616
+			for _, value := range values {
+				clonedReq.Header.Add(key, value)
+			}
+		}
+	}
+
+	return t.Origin.RoundTrip(clonedReq)
+}
+
+func (c *OpenAIClient) SetTools(tools []openai.Tool) {
+	c.tools = tools
+}
+
+func (c *OpenAIClient) Configure(config IAIConfig) error {
+	token := config.GetPassword()
+	cfg := openai.DefaultConfig(token)
+	orgId := config.GetOrganizationId()
+	proxyEndpoint := config.GetProxyEndpoint()
+
+	baseURL := config.GetBaseURL()
+	if baseURL != "" {
+		cfg.BaseURL = baseURL
+	}
+
+	transport := &http.Transport{}
+	if proxyEndpoint != "" {
+		proxyUrl, err := url.Parse(proxyEndpoint)
+		if err != nil {
+			return err
+		}
+		transport.Proxy = http.ProxyURL(proxyUrl)
+	}
+
+	if orgId != "" {
+		cfg.OrgID = orgId
+	}
+
+	customHeaders := config.GetCustomHeaders()
+	cfg.HTTPClient = &http.Client{
+		Transport: &OpenAIHeaderTransport{
+			Origin:  transport,
+			Headers: customHeaders,
+		},
+	}
+
+	client := openai.NewClientWithConfig(cfg)
+	if client == nil {
+		return errors.New("error creating OpenAI client")
+	}
+	c.client = client
+	c.model = config.GetModel()
+	c.temperature = config.GetTemperature()
+	c.topP = config.GetTopP()
+	c.maxHistory = config.GetMaxHistory()
+	c.memory = NewMemoryService()
+	return nil
+}
